fix: Pass API key to LiteLLM, more flexible model choices
mfcqi allows configuring an LLM vendor with key, so we should send
the key to LiteLLM if we have it -- otherwise, the user will need
to set an env var.

Meanwhile, the fixed choice of models is limited. Include the latest
GPT-5 models, but also, allow users to specify any model, and we'll
pass it through to the vendor. This way, mfcqi doesn't have to change
every time a new model comes out.